Metadata package
- Created new package called metadata
- separated different parts of metadata into different files
